Version 0.8
Former-commit-id: fc604c8b4e89505197f8fa015b63f02f8de19c62
Former-commit-id: 0f7f68fb2bb682378e97c5f8b0cb080f8b23abe2
Former-commit-id: 2fdc5ff4f360d973af62d50820ca1f966d2922ee

+++ App/Models/ProfileModel.php
@@ -30,19 +30,22 @@ class ProfileModel extends Model
         }
         $img = $this->upImg();
         if ($img) {
-// Deleting old photo before submitting the new one
-            $oldImg     = $user->img;
-            $oldImgPath = $this->app->file->toAvatar($oldImg);
-            unlink($oldImgPath);
-// Changing new image file permissions
-            $newImgPath = $this->app->file->toAvatar($img);
-            chmod($newImgPath, 0777);
-// Inserting the image filename in database
-            $this->db->data('img', $img);
+            if (!strstr($user->img, 'default/')) {
+                // Deleting old photo before submitting the new one
+                $oldImg     = $user->img;
+                $oldImgPath = $this->app->file->toAvatar($oldImg);
+                unlink($oldImgPath);
+                // Changing new image file permissions
+                $newImgPath = $this->app->file->toAvatar($img);
+                chmod($newImgPath, 0777);
+                // Inserting the image filename in database
+                $this->db->data('img', $img);
+            }
         }
         $this->db
                 ->data('name', $this->request->post('name'))
                 ->data('email', $this->request->post('email'))
+                ->data('bio', $this->request->post('bio'))
                 ->where('id = ?', $id)
                 ->update($this->table);
     }
